+++ src/mir/lower/effects.rs
@@ -52,7 +52,9 @@ impl Lowerer {
                 self.lower_block_stmts(body);
                 self.emit(InstKind::Void, Type::Void, *span)
             }
-            hir::Stmt::Together(name, body, span) => self.lower_together(*name, body, *span),
+            hir::Stmt::Together(name, body, errs, span) => {
+                self.lower_together(*name, body, errs, *span)
+            }
             hir::Stmt::UseLocal(_, _, _, _) => self.emit(InstKind::Void, Type::Void, Span::dummy()),
             hir::Stmt::GlobalStore(name, value, _span) => {
                 let val = self.lower_expr(value);