+++ src/typer/lower/resolve.rs
@@ -415,7 +415,7 @@ impl Typer {
             hir::Stmt::Transaction(block, _) => {
                 self.resolve_block(block);
             }
-            hir::Stmt::Together(_, block, _) => {
+            hir::Stmt::Together(_, block, _, _) => {
                 self.resolve_block(block);
             }
             hir::Stmt::ChannelClose(e, _) => self.resolve_expr(e),